+++ libbitcoinkernel-sys/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New `btck_chain_parameters_create_signet` for creating chain parameters for a custom signet from a user-provided challenge.
+
 ## [0.3.0] - 2026-05-20
 
 ### Added